correction

+++ 0x05-pointers_arrays_strings/8-print_array.c
@@ -8,7 +8,7 @@
 
 void print_array(int *a, int n)
 {
-	long long i;
+	long i;
 
 	for (i = 0; i < n; i++)
 	{
